bezier-utils, renderer 클래스 분리
