+++ src/lib/index.ts
@@ -1,4 +1,5 @@
 export * from './bezier-points';
+export * from './bezier-utils';
 export * from './canvas';
 export * from './config';
 export * from './dom';
