+++ src/renderer.ts
@@ -1,0 +1,151 @@
+import { DURATION, getBezierPoint, getInterpolatedPoint, type Point, STYLE } from './lib';
+
+export class CanvasRenderer {
+  private readonly staticCtx: CanvasRenderingContext2D;
+  private readonly dynamicCtx: CanvasRenderingContext2D;
+  private readonly width: number;
+  private readonly height: number;
+  private readonly pointColors: readonly string[];
+  private readonly finalPointColor: string;
+
+  constructor(
+    staticCtx: CanvasRenderingContext2D,
+    dynamicCtx: CanvasRenderingContext2D,
+    pointColors: readonly string[],
+    finalPointColor: string,
+  ) {
+    this.staticCtx = staticCtx;
+    this.dynamicCtx = dynamicCtx;
+    this.width = staticCtx.canvas.clientWidth;
+    this.height = staticCtx.canvas.clientHeight;
+    this.pointColors = pointColors;
+    this.finalPointColor = finalPointColor;
+  }
+
+  /** 베지에 곡선 가이드, 초기 조절점/레이블 같은 정적 요소 렌더링 */
+  public drawStaticLayer(points: Point[]): void {
+    const ctx = this.staticCtx;
+    this._clearLayer(ctx); // 베지에 곡선 차수(degree) 변경 시 가이드를 다시 그려야하므로 캔버스 초기화
+
+    const gridSize = this.width / STYLE.GRID_DIVISIONS;
+    // 좌상단, 우하단 경계는 제외하고 내부 영역에만 격자 추가
+    for (let x = gridSize; x < this.width; x += gridSize) {
+      this._drawLine(ctx, { x, y: 0 }, { x, y: this.height }, STYLE.GRID_COLOR); // 세로선(상 > 하 라인 생선)
+    }
+    for (let y = gridSize; y < this.height; y += gridSize) {
+      this._drawLine(ctx, { x: 0, y }, { x: this.width, y }, STYLE.GRID_COLOR); // 가로선(좌 > 우 라인 생성)
+    }
+
+    // ① 베지에 곡선 가이드
+    ctx.beginPath(); // 경로 시작
+    ctx.moveTo(points[0].x, points[0].y); // 시작점으로 이동
+    // t 증가값에 따라 곡선의 부드러움(품질)이 결정됨. 0.01은 총 100단계로 곡선을 그림
+    // 참고로 캔버스 네이티브 메서드(quadraticCurveTo, bezierCurveTo)로 2차, 3차 곡선을 그릴 수도 있음
+    for (let t = 0; t <= 1; t += DURATION.DELTA_T) {
+      const p = getBezierPoint(points, t);
+      ctx.lineTo(p.x, p.y); // 계산한 점까지 선 연결 (마지막 지점부터 이어서 연결)
+    }
+
+    ctx.strokeStyle = STYLE.GUIDE_COLOR;
+    ctx.lineWidth = STYLE.GUIDE_WIDTH;
+    ctx.stroke(); // 지정한 경로를 캔버스에 렌더링
+
+    // ② 조절점을 잇는 안내선 (점선)
+    ctx.setLineDash(STYLE.GUIDE_DASH); // 점선 모드로 변경. [점선 5px, 빈공간 5px] 형태로 반복
+    for (let i = 0; i < points.length - 1; i++) {
+      this._drawLine(ctx, points[i], points[i + 1], STYLE.GUIDE_COLOR);
+    }
+    ctx.setLineDash([]); // 점선 모드 해제 (실선으로 복원)
+
+    // ③ 조절점 및 레이블
+    points.forEach((p, i) => {
+      this._drawPoint(ctx, p, STYLE.CTRL_POINT_COLOR);
+      const offsetX = i % 2 === 0 ? -26 : 10;
+      this._drawLabel(ctx, p, `P${i}`, offsetX, -8);
+    });
+  }
+
+  /** 애니메이션의 각 프레임에서 t 값에 따라 변하는 동적 요소 렌더링 */
+  public drawDynamicLayer(points: Point[], t: number): void {
+    const ctx = this.dynamicCtx;
+    // 이전에 그렸던 모든 그래픽 요소 제거 (잔상처럼 보이는 현상 없애기 위해)
+    this._clearLayer(ctx);
+
+    let currentPoints = points;
+    let level = 0;
+
+    while (currentPoints.length > 1) {
+      const color = this.pointColors[level % this.pointColors.length];
+      const nextPoints: Point[] = [];
+
+      // 현재 레벨의 보간점 계산
+      for (let i = 0; i < currentPoints.length - 1; i++) {
+        const [p1, p2] = [currentPoints[i], currentPoints[i + 1]];
+        const interpolatedPoint = getInterpolatedPoint(p1, p2, t);
+        nextPoints.push(interpolatedPoint);
+        this._drawPoint(ctx, interpolatedPoint, color, STYLE.CTRL_POINT_RADIUS);
+      }
+
+      // 계산한 보간점 사이를 직선으로 연결
+      for (let i = 0; i < nextPoints.length - 1; i++) {
+        const [q1, q2] = [nextPoints[i], nextPoints[i + 1]];
+        this._drawLine(ctx, q1, q2, color, STYLE.INTERP_WIDTH);
+      }
+
+      currentPoints = nextPoints;
+      level++;
+    }
+
+    // 모든 보간이 끝나고 마지막 남은 점(베지에 곡선 위의 점) 표시
+    const finalPoint = currentPoints[0];
+    this._drawPoint(ctx, finalPoint, this.finalPointColor, STYLE.CTRL_POINT_RADIUS);
+    this._drawLabel(ctx, finalPoint, 'P', -20, 20);
+  }
+
+  private _clearLayer(ctx: CanvasRenderingContext2D) {
+    // clearRect() 메서드는 캔버스에서 특정 영역을 지울 때 사용
+    // clearRect(x, y, width, height) - (x,y) 좌표부터 width×height 크기만큼 지움
+    ctx.clearRect(0, 0, this.width, this.height);
+  }
+
+  private _drawLine(
+    ctx: CanvasRenderingContext2D,
+    start: Point,
+    end: Point,
+    color: string,
+    width: number = STYLE.BASE_WIDTH,
+  ): void {
+    ctx.beginPath(); // 새 경로 시작
+    ctx.moveTo(start.x, start.y); // 시작점 설정
+    ctx.lineTo(end.x, end.y); // 끝점까지 선 추가
+    ctx.strokeStyle = color; // 선 색상 설정
+    ctx.lineWidth = width; // 선 두께 설정
+    ctx.stroke(); // 정의된 경로에 따라 캔버스에 선 렌더링
+  }
+
+  private _drawPoint(
+    ctx: CanvasRenderingContext2D,
+    point: Point,
+    color: string,
+    radius: number = STYLE.BASE_POINT_RADIUS,
+  ): void {
+    ctx.beginPath();
+    // arc(x, y, radius, startAngle, endAngle, anticlockwise)
+    // 종료 각도를 나타내는 endAngle 인자는 라디언 단위이며 2 * Math.PI는 360도가 됨
+    ctx.arc(point.x, point.y, radius, 0, 2 * Math.PI);
+    ctx.fillStyle = color;
+    ctx.fill();
+  }
+
+  private _drawLabel(
+    ctx: CanvasRenderingContext2D,
+    point: Point,
+    text: string,
+    offsetX = 8,
+    offsetY = -8,
+  ): void {
+    ctx.font = STYLE.LABEL_FONT;
+    ctx.fillStyle = STYLE.LABEL_COLOR;
+    ctx.fillText(text, point.x + offsetX, point.y + offsetY);
+  }
+}
